lock/ops: remove circular import of keys operations
